warnings

+++ cmesh4.h
@@ -21,12 +21,12 @@ namespace cmesh4
   {
     static const uint64_t POINTS_IN_TRIANGLE = 3;
     SimpleMesh(){}
-    SimpleMesh(int a_vertNum, int a_indNum) { Resize(a_vertNum, a_indNum); }
+    SimpleMesh(size_t a_vertNum, size_t a_indNum) { Resize(a_vertNum, a_indNum); }
 
     inline size_t VerticesNum()  const { return vPos4f.size(); }
     inline size_t IndicesNum()   const { return indices.size();  }
     inline size_t TrianglesNum() const { return IndicesNum() / POINTS_IN_TRIANGLE;  }
-    inline void   Resize(uint32_t a_vertNum, uint32_t a_indNum)
+    inline void   Resize(size_t a_vertNum, size_t a_indNum)
     {
       vPos4f.resize(a_vertNum);
       vNorm4f.resize(a_vertNum);
